+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * use atexit to attempt to ensure main Lua state is destroyed so gcs run
 * setup a gc to kill spawned child processes
 * fix bug where HTTP client requests weren't notified on connection error
+* add a router message primitive
 
 
 ## 0.3.3 - 2016-05-21